changelog: Added CLI application and tests
main.go contains the CLI application for the COS Changelog project. It takes 2 build numbers as an input and outputs 2 JSON files containing the commit differences between them. This file is responsible for accepting user input, creating an authenticator object, passing arguments to the Changelog function in the Changelog package, converting the output to JSON, and writing the resulting JSON to a file. Changelog generation is handled by the Changelog package.

BUG=b/160901711
TEST=unittests, run local

Change-Id: Ic670f90328b4097c398c5ed7c9e8366f4a5ba864