changelog: Added querySize param and output struct to changelog pkg
A querySize parameter was added to the Changelog() function to accomodate the web application. Since we will not be displaying the entire commit list to users, it would be beneficial for performance if we had the ability to only query the amount of commits we want to display. The return output of Changelog was also changed from map[string][]*Commit to map[string]*RepoLog. RepoLog contains both a list of commits, the source and target committish, and an indicator of whether the list of commits is complete or not.

BUG=b/160901711
TEST=unittests, run local

Change-Id: Ieb62220fbb6e7e122107a9002c2ee4c09d37c231