changelog: Support multiple repos with same name
This CL addresses the issue with having multiple repositories with the same name (but different branches) in a manifest file. It adjusts the internal parsed manifest mapping to use repository name + repository path as the key, instead of only the repository name. It also tweaks the changelog output to use repository path as the repolog key, and stores the repository name as a value.

BUG=b/160901711
TEST=unittests, run local

Change-Id: I2306111040724977a03c8d540b6a5ea462b854e3