changelog-webapp: Added signout
This CL adds the ability to signout and changes the login flow. Users will only receive a login prompt as a first-time/signed out user, or after a month of inactivity. These users will be prompted to select an a account and grant permissions to the app. Otherwise, if the token is only expired, the login flow is completed automatically to receive a new access token.

BUG=b/160901711
TEST=run local

Change-Id: Ic296ebab1b9a6bc60b8ce3593f9aa56c8168fbbc