changelog: Remove repo-ID
Use repo path instead of repo-id (repo-name + repo-path) to match repositories for comparison. This allows repositories with 2 different names but the same path to be directly compared instead of treating them as separate repos and grouping them together in the webapp as additions and removals. Also fixed case where an empty changelog is displayed if an error occurs after calling additions().

BUG=b/160901711
TEST=unittests, run local

Change-Id: I888b34d54b297d59bf84cac550fcd4c8681c912b