Added config file to configure Profiler Tool
This CL adds a configuration file that acts as the user interface to the
Node Profiler tool. It also adds a function that parses user input from
the configuration file to a Logger Option structure. Changes needed in the
cloudlogger package are reflected in this CL.

BUG=b/196059500

Change-Id: I4d96500f4985b9191b8f29f0cfb48837dd0cc8a6
Reviewed-on: https://cos-review.googlesource.com/c/cos/tools/+/21010
Reviewed-by: Dexter Rivera <riverade@google.com>
Tested-by: Dexter Rivera <riverade@google.com>
Cloud-Build: GCB Service account <228075978874@cloudbuild.gserviceaccount.com>