Modify how MemCap collects util and saturation
This CL modifies how MemCap collects utilization such that
the only the percentage usage of main memory is captured. It also
modifies how saturation is collected such that a percentage of the
total Swap memory is used as a threshold. If total swap memory is 0,
a threshold is placed on utilization. This CL also modifies unit tests
accordingly.

Change-Id: I066d55c534c857572790598869ab3c9078708654
Reviewed-on: https://cos-review.googlesource.com/c/cos/tools/+/21110
Reviewed-by: Dexter Rivera <riverade@google.com>
Tested-by: Dexter Rivera <riverade@google.com>
Cloud-Build: GCB Service account <228075978874@cloudbuild.gserviceaccount.com>