Check for toolchain binaries explicitly
Currently, the check for toolchain existence is very broad, and passes
when the toolchain is not downloaded (because the kernel headers might
be downloaded, and the toolchain and kernel headers are downloaded to
the same location). Let's update this check to be more specific such
that it fails when the toolchain is not downloaded.

BUG=b/207547270
TEST=Run scripts/run_installer.sh with a test container with version
460.32.03.

Change-Id: I3a2b55c12abf046675484a25fa0fc58836e39b98
Reviewed-on: https://cos-review.googlesource.com/c/cos/tools/+/25560
Tested-by: Robert Kolchmeyer <rkolchmeyer@google.com>
Reviewed-by: Arnav Kansal <rnv@google.com>
Cloud-Build: GCB Service account <228075978874@cloudbuild.gserviceaccount.com>