cos-customizer: Select correct provisioner arch at runtime
We ship both provisioner_amd64 and provisioner_arm64 in cos-customizer,
and we update startup.yaml to select the correct one at runtime, based
on architecture.

Since arm64 ttys have a different name in /dev, we need to template that
into the customizer systemd service.

This change does change the behavior of cos-customizer. However, an
additional change is needed for users to actually use the arm64 code
path. So the behavioral changes shouldn't be noticeable to users right
now.

BUG=b/176990931
TEST=./run_tests

Change-Id: I8e0cd908126d128ee15f789ba72a61aee2c316e0
Reviewed-on: https://cos-review.googlesource.com/c/cos/tools/+/25901
Cloud-Build: GCB Service account <228075978874@cloudbuild.gserviceaccount.com>
Tested-by: Robert Kolchmeyer <rkolchmeyer@google.com>
Reviewed-by: Roy Yang <royyang@google.com>
Reviewed-by: Arnav Kansal <rnv@google.com>