cos_gpu_installer: add flag `-signature-url`
cos-gpu-intaller:v2 uses `-driver-version` to
find driver signature in cos-tools. But for
testing, `-nvidia-installer-url` is used so we
cannot pass `-driver-version` and we cannot
release signature for testing to cos-tools.
So this CL adds a flag `-signature-url` to find
signature at a URL in test mode.

`-signature-url` can only be used together with
`-nvidia-installer-url` and `-test`.

BUG=b/161566626

Change-Id: Ibb55004c6c8f19b1cbb82ec755df10a27d86428f
Reviewed-on: https://cos-review.googlesource.com/c/cos/tools/+/30400
Cloud-Build: GCB Service account <228075978874@cloudbuild.gserviceaccount.com>
Reviewed-by: Robert Kolchmeyer <rkolchmeyer@google.com>
Tested-by: He Gao <hegao@google.com>