cos_gpu_installer: add --prepare-build-tools flag which populates toolchain cache
This flag can be used to populate the build tools cache. This includes
downloading and installing the toolchain and the kernel headers which
are a pre-requisite to the nvidia driver installer. Using this flag will NOT
install the driver and this will not depend on existence of a GPU
attached to the instance.

BUG=b/175236375
TEST=manual invokation of installer with --prepare-build-tools

Change-Id: I7414efdb8ee9ff380358fe3e6cd86f3c232778ab
Reviewed-on: https://cos-review.googlesource.com/c/cos/tools/+/30822
Cloud-Build: GCB Service account <228075978874@cloudbuild.gserviceaccount.com>
Reviewed-by: He Gao <hegao@google.com>
Tested-by: Arnav Kansal <rnv@google.com>