cos_gpu_insatller: do not load public key or check release track in test mode
In test mode, platform key is used which is already in
a derived test image. There is no need to load the public
key and it cannot be loaded to the secondary trusted keyring.

All prerelease builds are in dev-channel. For testing we
don't need to check release track.

Change-Id: Id8e686524bd4c9dc106d268500ba3ea43a5a9f1b
Reviewed-on: https://cos-review.googlesource.com/c/cos/tools/+/31180
Cloud-Build: GCB Service account <228075978874@cloudbuild.gserviceaccount.com>
Reviewed-by: Arnav Kansal <rnv@google.com>
Tested-by: He Gao <hegao@google.com>