cos-gpu-installer-v1: forcefully umount /usr/lib/x86_64-linux-gnu
When we upgraded the debian base image in cos-gpu-installer, the umount
for /usr/lib/x86_64-linux-gnu started to fail because target was busy.
This is blocking the release of new gpu installer in cos-customizer.
This issue is not present in cos-gpu-installer-v2.

BUG=b/225958311
TEST=ran cos-gpu-installer-v1 on an image
RELEASE_NOTES=None

Change-Id: I97495fcee24690769d7c5ba0aa88f19689d1abd8
Reviewed-on: https://cos-review.googlesource.com/c/cos/tools/+/31761
Cloud-Build: GCB Service account <228075978874@cloudbuild.gserviceaccount.com>
Reviewed-by: Arnav Kansal <rnv@google.com>
Tested-by: Rayan Dasoriya <dasoriya@google.com>