COS tools: add .gitignore and a few testing instructions
This clarifies that `bazel` and `mtools` are required to run tests for
the COS tools repo, and that a `python` binary must be available on the
`$PATH`. It also mentions that ~10 K80 GPUs in us-west1 are needed to
successfully execute `./run_tests.sh`; my experience says that 5 is not
enough, but I'm told that 10 is.

The local Docker-build instructions for cos-customizer are also updated
to clarify that it refers to `//:cos_customizer`, not the target in the
same directory as the instructions,
`//src/cmd/cos_customizer:cos_customizer`.

Added a `.gitignore` file, as well, to exclude `bazel-*` directories at
the top level of the project from Git.

BUG=b/229419550
TEST=none; just doc updates

Change-Id: I0dc9065e76f78e5ffe91cdf0aaec0fc4619fa227
Reviewed-on: https://cos-review.googlesource.com/c/cos/tools/+/32100
Reviewed-by: Robert Kolchmeyer <rkolchmeyer@google.com>
Tested-by: Robert Kolchmeyer <rkolchmeyer@google.com>
Reviewed-by: Arnav Kansal <rnv@google.com>
Cloud-Build: GCB Service account <228075978874@cloudbuild.gserviceaccount.com>