Fix arch-specific paths in the artifact archival code
After switching to the ChromiumOS-compatible architecture names
some of the paths in the kernel tree were rendered invalid because
of the mismatch between the CrOS arch name (x86_64) and the kernel
tree one (x86). As a fix do additional translation of the arch name
to fix the paths.

Change-Id: I4aceb5de53861460a9f0ce357c1a4e939d9ddd95
Reviewed-on: https://cos-review.googlesource.com/c/cos/tools/+/32340
Cloud-Build: GCB Service account <228075978874@cloudbuild.gserviceaccount.com>
Reviewed-by: Meena Shanmugam <meenashanmugam@google.com>
Tested-by: Oleksandr Tymoshenko <ovt@google.com>