Add kselftest-builder container for x86 and ARM64
Kernel's sefltest depends on a number of userland libraries
COS does not provide. For this reason Ubuntu is used as a base
image and should be used as a run-time test userland.

kselftest also does not fully support cross-compilation so
the builder needs ARM64 version that can be used with linuxkit/binfmt
on x86 worker VM.

Change-Id: I00dccad76ec83d12ffb601850e2e48f801ab7bc5
Reviewed-on: https://cos-review.googlesource.com/c/cos/tools/+/34940
Reviewed-by: Oleksandr Tymoshenko <ovt@google.com>
Reviewed-by: Arnav Kansal <rnv@google.com>
Tested-by: Oleksandr Tymoshenko <ovt@google.com>
Cloud-Build: GCB Service account <228075978874@cloudbuild.gserviceaccount.com>