tools: Adding documentation to make the implementation details of the customizer a bit clearer.
Since there's quite a bit of integration involved in the customizer, I added a few docs to try
and clear things up about how the customizer works. This commit adds two README files to
the provisioner and preloader directories as well as a brief snippet in the finish-image-build step
that explains the process of how resources are created, modified, and then cleaned up.

BUG=b/240175057,b/172925856
TEST=None

Change-Id: I90630e11e7117909366fd6ce104029ebde013a9c
Reviewed-on: https://cos-review.googlesource.com/c/cos/tools/+/36129
Reviewed-by: Robert Kolchmeyer <rkolchmeyer@google.com>
Cloud-Build: GCB Service account <228075978874@cloudbuild.gserviceaccount.com>
Tested-by: Nobel Barakat <nobelbarakat@google.com>