Add ProcessConfigs method for GPU precompilation
This method include the implementation of finding all the configs that
have not been completely processed. For each such config it invokes the
builder methods which create the -custom.run self extracting archive and
also uploads this precompiled driver to the output directory as specified in the config.

BUG=b/238215395

Change-Id: I7f787b6520fd76a0d0b3424fbe1b7f4b4c969f11
Reviewed-on: https://cos-review.googlesource.com/c/cos/tools/+/37747
Cloud-Build: GCB Service account <228075978874@cloudbuild.gserviceaccount.com>
Reviewed-by: He Gao <hegao@google.com>
Tested-by: Arnav Kansal <rnv@google.com>