cos_gpu_installer: pass driver version to firmware installation
Nvidia module automatically loads firmware from a very specific path
inside the firmware search paths.

Specifically it will query the paths + "nvidia/<driver-version>/" dir.

For the driver installation that takes the route of
--nvidia-installer-url flag the driver-version flag is missing.

This becomes an issue because now the firmware is installed in the dir
"nvidia/" instead of the "nvidia/<driver-version>" directory.

The .manifest file of all installers contain the driver version as part
of the identification text. This change tries to find a driver version
on the manifest file and use that if the driver version is not supplied.

TEST plan:
Ran gpu_e2e_test with this installer to check -nvidia-installer-url path
and also ran gpu driver installation on cos-lts builds to check
-driver-version path.

Change-Id: I092dcef119fcbc291452a8e74d4b42e6348f4153
Reviewed-on: https://cos-review.googlesource.com/c/cos/tools/+/41927
Reviewed-by: Meena Shanmugam <meenashanmugam@google.com>
Tested-by: Arnav Kansal <rnv@google.com>
Cloud-Build: GCB Service account <228075978874@cloudbuild.gserviceaccount.com>