Fix Containers losing access to GPUs with error: "Failed to initialize NVML: Unknown Error"
Containerized GPU workloads may suddenly lose access to their GPUs. This
situation occurs when systemd is used to manage the cgroups of the
container and it is triggered to reload any Unit files that have
references to NVIDIA GPUs (e.g. with something as simple as a systemctl
daemon-reload).

The issue originates from the fact that recent versions of runc require
that symlinks be present under /dev/char to any device nodes being
injected into a container. Unfortunately, these symlinks are not present
for NVIDIA devices, and the NVIDIA GPU driver does not (currently)
provide a means for them to be created automatically.

Manually create the symlinks in /dev/char for all possible NVIDIA device
nodes required for using GPUs in containers using the nvidia-ctk
utility.

BUG=b/267345708

Change-Id: I8d6478aec766f51fe3c55bf9940047fc2eadf48e
Reviewed-on: https://cos-review.googlesource.com/c/cos/tools/+/43827
Reviewed-by: Meena Shanmugam <meenashanmugam@google.com>
Tested-by: Arnav Kansal <rnv@google.com>
Cloud-Build: GCB Service account <228075978874@cloudbuild.gserviceaccount.com>
Reviewed-by: Robert Kolchmeyer <rkolchmeyer@google.com>