cos-gpu-installer: change permission of NVIDIA install dir to 755
For the OSS(prebuilt) driver modules, we extract the prebuilt drivers
(kernel modules, firmware) to the install dir, before loading the GPU drivers.

The install dir needs to have permission 755 for non root users to be
able to use the drivers.

BUG=b/295358831
TEST=install latest drivers on COS
compared permissions of drivers on older milestones with default driver
to the latest drivers installed with this change.
https://diff.googleplex.com/#key=RJbGZPya4FwC

Change-Id: I43a3387d7df2ab5d8a3db8f3996947e0eee2822b
Reviewed-on: https://cos-review.googlesource.com/c/cos/tools/+/54487
Cloud-Build: GCB Service account <228075978874@cloudbuild.gserviceaccount.com>
Reviewed-by: Robert Kolchmeyer <rkolchmeyer@google.com>
Tested-by: Arnav Kansal <rnv@google.com>